requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
rich>=13.0.0
selenium>=4.15.0
pyautogui>=0.9.54
opencv-python>=4.8.0
pytesseract>=0.3.10
PyPDF2>=3.0.1
python-docx>=0.8.11
pillow>=10.0.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0